Moved global endpoints to linode client
