+++ test/objects/networking_test.py
@@ -55,65 +55,3 @@ class NetworkingTest(ClientBaseCase):
             self.assertEqual(result["outbound"], [])
             self.assertEqual(result["inbound_policy"], "DROP")
             self.assertEqual(result["outbound_policy"], "DROP")
-
-    def test_ip_addresses_share(self):
-        """
-        Tests that you can submit a correct ip addresses share api request.
-        """
-
-        ip = IPAddress(self.client, "192.0.2.1", {})
-
-        with self.mock_post({}) as m:
-            ip.ip_addresses_share(["192.0.2.1"], 123)
-            self.assertEqual(m.call_url, "/networking/ips/share")
-            self.assertEqual(m.call_data["ips"], ["192.0.2.1"])
-            self.assertEqual(m.call_data["linode_id"], 123)
-
-        with self.mock_post({}) as m:
-            ip.ip_addresses_share([ip], 123)
-            self.assertEqual(m.call_url, "/networking/ips/share")
-            self.assertEqual(m.call_data["ips"], ["192.0.2.1"])
-            self.assertEqual(m.call_data["linode_id"], 123)
-
-    def test_ip_addresses_assign(self):
-        """
-        Tests that you can submit a correct ip addresses assign api request.
-        """
-
-        ip = IPAddress(self.client, "192.0.2.1", {})
-
-        with self.mock_post({}) as m:
-            ip.ip_addresses_assign(
-                {"assignments": [{"address": "192.0.2.1", "linode_id": 123}]},
-                "us-east",
-            )
-            self.assertEqual(m.call_url, "/networking/ips/assign")
-            self.assertEqual(
-                m.call_data["assignments"],
-                {"assignments": [{"address": "192.0.2.1", "linode_id": 123}]},
-            )
-            self.assertEqual(m.call_data["region"], "us-east")
-
-    def test_ip_ranges_list(self):
-        """
-        Tests that you can submit a correct ip ranges list api request.
-        """
-
-        ipv6Range = IPv6Range(self.client, "2600:3c01::")
-        ipv6Range._api_get()
-
-        with self.mock_post("/networking/ipv6/ranges") as m:
-            result = ipv6Range.ip_ranges_list()
-            self.assertEqual(m.call_url, "/networking/ipv6/ranges")
-            self.assertEqual(len(result), 1)
-
-    def test_ip_range_delete(self):
-        """
-        Tests that you can submit a correct ip range delete api request.
-        """
-
-        ipv6Range = IPv6Range(self.client, "2600:3c01::")
-
-        with self.mock_delete() as m:
-            ipv6Range.ip_range_delete()
-            self.assertEqual(m.call_url, "/networking/ipv6/ranges/2600:3c01::")
